[Chore 159799611] Separate user code from auth code :art:
