+++ server/middleware/validation/index.js
@@ -1,7 +1,9 @@
 import Entry from './modules/entry';
 import Auth from './modules/auth';
+import User from './modules/user';
 
 module.exports = {
   Entry,
   Auth,
+  User,
 };
