+++ server/router/index.js
@@ -1,8 +1,10 @@
 import entries from './modules/entries';
 import auth from './modules/auth';
+import users from './modules/users';
 
 
 module.exports = {
   entries,
   auth,
+  users,
 };
