+++ server/server.js
@@ -36,6 +36,7 @@ app.use((req, res, next) => {
 
 app.use('/api/v1/entries', router.entries);
 app.use('/api/v1/auth', router.auth);
+app.use('/api/v1/user', router.users);
 app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerDoc));
 
 // when there is no fitting route set error and run the next func
